move reader functions to the readers.py file
