+++ src/biome/data/sources/__init__.py
@@ -1,5 +1,3 @@
-from .helpers import read_dataset
-
+from .datasource import DataSource
 from .example_preparator import SOURCE_FIELD, RESERVED_FIELD_PREFIX
-
 from .utils import ID, RESOURCE
